Pull updateTransactionHistory into history file and change name of the file
